+++ cogs/core.py
@@ -1,6 +1,5 @@
 from discord.ext import commands
-from discord.utils import get
-import discord, traceback, asyncio, json, datetime
+import discord, traceback, json, datetime
 BOTSETFILE = "ext/bot_settings.json"
 SETFILE = "data/settings.json"
 
@@ -15,15 +14,14 @@ class Core(commands.Cog):
         except Exception: pass
         if args:
             command = self.bot.get_command(args)
-            if not command: return await ctx.send('Command not found, please try again.')
+            if not command or command.hidden: return await ctx.send('Command not found, please try again.')
             e = discord.Embed(title='Command `/' + ((' '.join([p.name for p in command.parents]) + ' ' + command.name) if (command.parents) else (command.name)) + '`', description=(command.description or "<no description>"))
             e.add_field(name='Objective',   value=command.help)
             e.add_field(name='Usage',       value=command.usage)
             e.add_field(name='Cog',         value="No cogs" if not command.cog else command.cog.qualified_name)
             if hasattr(command, 'commands'):    # it is a group
                 e.add_field(name='Sub-Commands', value=', '.join([cmd.name for cmd in command.commands]))
-            msg = await ctx.send(embed=e)
-            names = [(None if cmd.hidden else cmd.name) for cmd in self.bot.commands]  # loop over all commands, if not hidden, append its string name
+            await ctx.send(embed=e)
             return
 
         all_cmds = self.bot.commands